calculatePrice() from Order moved to Cart
