+++ src/Cart.java
@@ -35,4 +35,22 @@ public class Cart {
     public ArrayList<CartItem> getItems() {
         return items;
     }
+
+    public double calculatePrice(String subscription) {
+        double totalPrice = 0.0;
+
+        for (CartItem item : items) {
+            totalPrice += item.getTotalPrice();
+        }
+
+        if (subscription == "gold") {
+            totalPrice *= 0.15; // 15% discount for prime members
+        } else if (subscription == "platinum") {
+            totalPrice *= 0.10; // 10% discount for platinum members
+        } else if (subscription == "silver") {
+            totalPrice *= 0.05; // 5% discount for silver members
+        } 
+
+        return totalPrice;
+    }
 }
